add ui design navBar
