+++ hamsa_lomi/lib/app.dart
@@ -13,7 +13,7 @@ class App extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: 'Hamsa Lomi',
       theme: HamsaTheme.lightTheme,
-      home: MyHomePage(),
+      home: BottomNav(),
     );
   }
 }
