+++ hamsa_lomi/lib/presentation/page/test_bottom_bar.dart
@@ -1,270 +1,66 @@
 import 'package:flutter/material.dart';
+import '../constants/app_assets_constant.dart';
+import '../share_widgets/appbar_widget.dart';
+import '../share_widgets/bottombar_widget.dart';
 // import 'package:flutter_animated_bottom/custom_animated_bottom_bar.dart';
 
-class MyHomePage extends StatefulWidget {
+class BottomNav extends StatefulWidget {
 
   @override
-  _MyHomePageState createState() => _MyHomePageState();
+  _BottomNavState createState() => _BottomNavState();
 }
 
-class _MyHomePageState extends State<MyHomePage> {
+class _BottomNavState extends State<BottomNav> {
   int _currentIndex = 0;
 
   final _inactiveColor = Colors.grey;
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          automaticallyImplyLeading: false,
-          title: Text("Custom Animated Bottom Navigation Bar"),
-          backgroundColor: Colors.green[200],
+      backgroundColor: HamsaColors.lightBackground,
+        appBar: HamsaAppBar(
+          withLogo: true,
+          appBarText: Text('Testing bottom nav'),
         ),
-        body: getBody(),
         bottomNavigationBar: _buildBottomBar()
     );
   }
 
-  Widget _buildBottomBar(){
-    return CustomAnimatedBottomBar(
-      containerHeight: 70,
-      backgroundColor: Colors.black,
+  Widget _buildBottomBar() {
+    return HamsaBottomBar(
+      containerHeight: 75,
+      backgroundColor: HamsaColors.lightBackground,
       selectedIndex: _currentIndex,
       showElevation: true,
-      itemCornerRadius: 24,
+      itemCornerRadius: 15,
       curve: Curves.easeIn,
       onItemSelected: (index) => setState(() => _currentIndex = index),
       items: <BottomNavyBarItem>[
         BottomNavyBarItem(
-          icon: Icon(Icons.apps),
+          icon: Icon(Icons.home),
           title: Text('Home'),
-          activeColor: Colors.green,
+          activeColor: HamsaColors.primaryColor,
           inactiveColor: _inactiveColor,
           textAlign: TextAlign.center,
         ),
         BottomNavyBarItem(
-          icon: Icon(Icons.people),
-          title: Text('Users'),
-          activeColor: Colors.purpleAccent,
+          icon: Icon(Icons.apps),
+          title: Text('FUNDRAISERS'),
+          activeColor: HamsaColors.primaryColor,
           inactiveColor: _inactiveColor,
           textAlign: TextAlign.center,
         ),
         BottomNavyBarItem(
-          icon: Icon(Icons.message),
+          icon: Icon(Icons.person),
           title: Text(
-            'Messages ',
+            'PROFILE ',
           ),
-          activeColor: Colors.pink,
-          inactiveColor: _inactiveColor,
-          textAlign: TextAlign.center,
-        ),
-        BottomNavyBarItem(
-          icon: Icon(Icons.settings),
-          title: Text('Settings'),
-          activeColor: Colors.blue,
+          activeColor: HamsaColors.primaryColor,
           inactiveColor: _inactiveColor,
           textAlign: TextAlign.center,
         ),
       ],
     );
   }
-
-
-  Widget getBody() {
-    List<Widget> pages = [
-      Container(
-        alignment: Alignment.center,
-        child: Text("Home",style: TextStyle(fontSize: 25,fontWeight: FontWeight.bold),),
-      ),
-      Container(
-        alignment: Alignment.center,
-        child: Text("Users",style: TextStyle(fontSize: 25,fontWeight: FontWeight.bold),),
-      ),
-      Container(
-        alignment: Alignment.center,
-        child: Text("Messages",style: TextStyle(fontSize: 25,fontWeight: FontWeight.bold),),
-      ),
-      Container(
-        alignment: Alignment.center,
-        child: Text("Settings",style: TextStyle(fontSize: 25,fontWeight: FontWeight.bold),),
-      ),
-    ];
-    return IndexedStack(
-      index: _currentIndex,
-      children: pages,
-    );
-  }
-
-
-}
-
-class CustomAnimatedBottomBar extends StatelessWidget {
-
-  CustomAnimatedBottomBar({
-    Key? key,
-    this.selectedIndex = 0,
-    this.showElevation = true,
-    this.iconSize = 24,
-    this.backgroundColor,
-    this.itemCornerRadius = 50,
-    this.containerHeight = 56,
-    this.animationDuration = const Duration(milliseconds: 270),
-    this.mainAxisAlignment = MainAxisAlignment.spaceBetween,
-    required this.items,
-    required this.onItemSelected,
-    this.curve = Curves.linear,
-  }) : assert(items.length >= 2 && items.length <= 5),
-        super(key: key);
-
-  final int selectedIndex;
-  final double iconSize;
-  final Color? backgroundColor;
-  final bool showElevation;
-  final Duration animationDuration;
-  final List<BottomNavyBarItem> items;
-  final ValueChanged<int> onItemSelected;
-  final MainAxisAlignment mainAxisAlignment;
-  final double itemCornerRadius;
-  final double containerHeight;
-  final Curve curve;
-
-  @override
-  Widget build(BuildContext context) {
-    final bgColor = backgroundColor ?? Theme.of(context).bottomAppBarColor;
-
-    return Container(
-      decoration: BoxDecoration(
-        color: bgColor,
-        boxShadow: [
-          if (showElevation)
-            const BoxShadow(
-              color: Colors.black12,
-              blurRadius: 2,
-            ),
-        ],
-      ),
-      child: SafeArea(
-        child: Container(
-          width: double.infinity,
-          height: containerHeight,
-          padding: const EdgeInsets.symmetric(vertical: 6, horizontal: 8),
-          child: Row(
-            mainAxisAlignment: mainAxisAlignment,
-            children: items.map((item) {
-              var index = items.indexOf(item);
-              return GestureDetector(
-                onTap: () => onItemSelected(index),
-                child: _ItemWidget(
-                  item: item,
-                  iconSize: iconSize,
-                  isSelected: index == selectedIndex,
-                  backgroundColor: bgColor,
-                  itemCornerRadius: itemCornerRadius,
-                  animationDuration: animationDuration,
-                  curve: curve,
-                ),
-              );
-            }).toList(),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ItemWidget extends StatelessWidget {
-  final double iconSize;
-  final bool isSelected;
-  final BottomNavyBarItem item;
-  final Color backgroundColor;
-  final double itemCornerRadius;
-  final Duration animationDuration;
-  final Curve curve;
-
-  const _ItemWidget({
-    Key? key,
-    required this.item,
-    required this.isSelected,
-    required this.backgroundColor,
-    required this.animationDuration,
-    required this.itemCornerRadius,
-    required this.iconSize,
-    this.curve = Curves.linear,
-  })  : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Semantics(
-      container: true,
-      selected: isSelected,
-      child: AnimatedContainer(
-        width: isSelected ? 130 : 50,
-        height: double.maxFinite,
-        duration: animationDuration,
-        curve: curve,
-        decoration: BoxDecoration(
-          color:
-          isSelected ? item.activeColor.withOpacity(0.2) : backgroundColor,
-          borderRadius: BorderRadius.circular(itemCornerRadius),
-        ),
-        child: SingleChildScrollView(
-          scrollDirection: Axis.horizontal,
-          physics: NeverScrollableScrollPhysics(),
-          child: Container(
-            width: isSelected ? 130 : 50,
-            padding: EdgeInsets.symmetric(horizontal: 8),
-            child: Row(
-              mainAxisSize: MainAxisSize.max,
-              mainAxisAlignment: MainAxisAlignment.start,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: <Widget>[
-                IconTheme(
-                  data: IconThemeData(
-                    size: iconSize,
-                    color: isSelected
-                        ? item.activeColor.withOpacity(1)
-                        : item.inactiveColor == null
-                        ? item.activeColor
-                        : item.inactiveColor,
-                  ),
-                  child: item.icon,
-                ),
-                if (isSelected)
-                  Expanded(
-                    child: Container(
-                      padding: EdgeInsets.symmetric(horizontal: 4),
-                      child: DefaultTextStyle.merge(
-                        style: TextStyle(
-                          color: item.activeColor,
-                          fontWeight: FontWeight.bold,
-                        ),
-                        maxLines: 1,
-                        textAlign: item.textAlign,
-                        child: item.title,
-                      ),
-                    ),
-                  ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-class BottomNavyBarItem {
-
-  BottomNavyBarItem({
-    required this.icon,
-    required this.title,
-    this.activeColor = Colors.blue,
-    this.textAlign,
-    this.inactiveColor,
-  });
-
-  final Widget icon;
-  final Widget title;
-  final Color activeColor;
-  final Color? inactiveColor;
-  final TextAlign? textAlign;
-
 }
