+++ hamsa_lomi/lib/presentation/share_widgets/bottombar_widget.dart
@@ -1,11 +1,168 @@
 // Flutter imports:
 import 'package:flutter/material.dart';
+import 'package:hamsa_lomi/presentation/constants/app_assets_constant.dart';
 
 class HamsaBottomBar extends StatelessWidget {
-  const HamsaBottomBar({Key? key}) : super(key: key);
+
+  HamsaBottomBar({
+    Key? key,
+    this.selectedIndex = 0,
+    this.showElevation = true,
+    this.iconSize = 20,
+    this.backgroundColor,
+    this.itemCornerRadius = 50,
+    this.containerHeight = 56,
+    this.animationDuration = const Duration(milliseconds: 270),
+    this.mainAxisAlignment = MainAxisAlignment.spaceBetween,
+    required this.items,
+    required this.onItemSelected,
+    this.curve = Curves.linear,
+  }) : assert(items.length >= 2 && items.length <= 5),
+        super(key: key);
+
+  final int selectedIndex;
+  final double iconSize;
+  final Color? backgroundColor;
+  final bool showElevation;
+  final Duration animationDuration;
+  final List<BottomNavyBarItem> items;
+  final ValueChanged<int> onItemSelected;
+  final MainAxisAlignment mainAxisAlignment;
+  final double itemCornerRadius;
+  final double containerHeight;
+  final Curve curve;
 
   @override
   Widget build(BuildContext context) {
-    return BottomAppBar();
+    final bgColor = backgroundColor ?? Theme.of(context).bottomAppBarColor;
+
+    return Container(
+      decoration: BoxDecoration(
+        color: bgColor,
+      ),
+      child: SafeArea(
+        child: Container(
+          width: double.infinity,
+          height: containerHeight,
+          padding: const EdgeInsets.symmetric(vertical: 5, horizontal: 8),
+          child: Row(
+            mainAxisAlignment: mainAxisAlignment,
+            children: items.map((item) {
+              var index = items.indexOf(item);
+              return GestureDetector(
+                onTap: () => onItemSelected(index),
+                child: _ItemWidget(
+                  item: item,
+                  iconSize: iconSize,
+                  isSelected: index == selectedIndex,
+                  backgroundColor: bgColor,
+                  itemCornerRadius: itemCornerRadius,
+                  animationDuration: animationDuration,
+                  curve: curve,
+                ),
+              );
+            }).toList(),
+          ),
+        ),
+      ),
+    );
   }
+}
+
+class _ItemWidget extends StatelessWidget {
+  final double iconSize;
+  final bool isSelected;
+  final BottomNavyBarItem item;
+  final Color backgroundColor;
+  final double itemCornerRadius;
+  final Duration animationDuration;
+  final Curve curve;
+
+  const _ItemWidget({
+    Key? key,
+    required this.item,
+    required this.isSelected,
+    required this.backgroundColor,
+    required this.animationDuration,
+    required this.itemCornerRadius,
+    required this.iconSize,
+    this.curve = Curves.linear,
+  })  : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Semantics(
+      container: true,
+      selected: isSelected,
+      child: AnimatedContainer(
+        width: isSelected ? 130 : 50,
+        height: double.maxFinite,
+        duration: animationDuration,
+        curve: curve,
+        decoration: BoxDecoration(
+          color:
+          isSelected ? item.activeColor.withOpacity(0.2) : backgroundColor,
+          borderRadius: BorderRadius.circular(itemCornerRadius),
+        ),
+        child: SingleChildScrollView(
+          scrollDirection: Axis.horizontal,
+          physics: NeverScrollableScrollPhysics(),
+          child: Container(
+            width: isSelected ? 130 : 50,
+            padding: EdgeInsets.symmetric(horizontal: 8),
+            child: Row(
+              mainAxisSize: MainAxisSize.max,
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: <Widget>[
+                IconTheme(
+                  data: IconThemeData(
+                    size: iconSize,
+                    color: isSelected
+                        ? item.activeColor.withOpacity(1)
+                        : item.inactiveColor == null
+                        ? item.activeColor
+                        : item.inactiveColor,
+                  ),
+                  child: item.icon,
+                ),
+                if (isSelected)
+                  Expanded(
+                    child: Container(
+                      padding: EdgeInsets.symmetric(horizontal: 4),
+                      child: DefaultTextStyle.merge(
+                        style: TextStyle(
+                          color: item.activeColor,
+                          fontWeight: FontWeight.bold,
+                        ),
+                        maxLines: 1,
+                        textAlign: item.textAlign,
+                        child: item.title,
+                      ),
+                    ),
+                  ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
+class BottomNavyBarItem {
+
+  BottomNavyBarItem({
+    required this.icon,
+    required this.title,
+    this.activeColor = HamsaColors.primaryColor,
+    this.textAlign,
+    this.inactiveColor,
+  });
+
+  final Widget icon;
+  final Widget title;
+  final Color activeColor;
+  final Color? inactiveColor;
+  final TextAlign? textAlign;
+
 }
